\name{JASPAR2014SitesSeqs}
\alias{JASPAR2014SitesSeqs}
\docType{data}
\title{
  Sites sequences
}
\description{
  A list of DNAStringSet storing transcription factor binding sites sequences 
  from JASPAR 2014 release with JASPAR IDs as names
}
\source{
  \url{http://jaspar.binf.ku.dk/html/DOWNLOAD/sites/}
}
\examples{
  \dontrun{
    library(JASPAR2014)
    JASPAR2014SitesSeqs
  }
}
\keyword{datasets}
