\name{JASPAR2014-class}
\Rdversion{1.1}
\docType{class}
\alias{JASPAR2014-class}
\alias{JASPAR2014}

\title{JASPAR2014 object}
\description{
  The JASPAR2014 object class is a thin class for storing the path of JASPAR2014 style SQLite file. 
}

\section{Slots}{
  \describe{
    \item{\code{db}:}{Object of class \code{"character"}: a character string of the path of SQLite file.}
  }
}

\references{
%%  ~~put references to the literature/web site here~~
}
\author{
  Ge Tan
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
  \link{JASPAR2014SiteSeqs},
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
  library(JASPAR2014)
  JASPAR2014
}
\keyword{classes}
